requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.0.1
numpy==1.24.3
tensorflow-lite==2.13.0
pydantic==2.4.2

